requirements.txt
asyncio-mqtt==0.16.2
asyncpg==0.28.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
chirpstack-api==4.4.0
google-api-core==2.11.1
google-auth==2.21.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-tools==1.56.0
idna==3.4
paho-mqtt==1.6.1
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
ujson==5.8.0
urllib3==1.26.16